finalized
finalized the flow.

Sometimes like the 5th number will generate a lower number than the 3rd of 4th number. The order of numbers might differ because of the sort() routine. But none the less it is still with in the range based from the history of drawings of powerball.

goodluck and enjoy.